Edit : Data type modification

+++ BattleGame/Assets/Scripts/Core/Character/Enemy/GroupRush/GroupRushAttribute.cs
@@ -11,7 +11,7 @@ namespace Core.Character.Enemy.GroupRush
 
         public float EnemySpeed { get => enemySpeed; set => enemySpeed = value; }
         public int EnemyStamina { get => enemyStamina; set => enemyStamina = value; }
-        public int SpawnInterbal { get => spawnInterbal; set => spawnInterbal = value; }
+        public float SpawnInterbal { get => spawnInterbal; set => spawnInterbal = value; }
 
 
     }
